refactor: 추가 쪼개기
