+++ config/packages.mk
@@ -14,9 +14,9 @@
 PKGS = engine parser raycast renderer
 
 engineV = collision door init keyinput movement movement_impl run
-parserV = check__cubfile check__cubid check__map check__map2\
+parserV = check__cubfile check__cubid check__map check__map__format\
 		check__map__index check__parser cubid fluidfill map map__raw\
-		parser rgb
+		parser rgb sprite world
 raycastV = floor raycast texture wall
 rendererV = camera dda__dist dda__step deinit draw image init load_tex\
 			minimap util
