+++ src/parser/sprite.c
@@ -1,0 +1,19 @@
+#include "types__parser.h"
+#include "types__renderer.h"
+
+// static int	count_num_sprites(t_world *world);
+
+// void	world__init__sprites(t_world *world)
+// {
+// 	t_i64vec		it;
+
+// 	it.y = -1;
+// 	while (++it.y < world->world_height)
+// 	{
+// 		it.x = -1;
+// 		while (++it.x < world->world_width)
+// 		{
+
+// 		}
+// 	}
+// }
