+++ src/renderer/deinit.c
@@ -21,6 +21,9 @@ void	world__deinit(t_world *this)
 	while (++i < this->world_height)
 		free(this->worldmap[i]);
 	free(this->worldmap);
+	// free(this->sprites);
+	// free(this->sprite_orders);
+	// free(this->sprite_distances);
 }
 
 void	renderer__deinit(t_renderer *this)
